set up server.js
